+++ Source/Modules/Catalog/Math/Gapotchenko.FX.Math.Graphs/Graph`1.Degree.cs
@@ -13,21 +13,37 @@ partial class Graph<TVertex>
     /// <inheritdoc />
     public int GetVertexIndegree(TVertex vertex)
     {
+        if (IsDirected)
+            return GetVertexIndegreeCore(vertex);
+        else
+            return GetVertexDegree(vertex);
+    }
+
+    /// <inheritdoc />
+    public int GetVertexOutdegree(TVertex vertex)
+    {
+        if (IsDirected)
+            return GetVertexOutdegreeCore(vertex);
+        else
+            return GetVertexDegree(vertex);
+    }
+
+    /// <inheritdoc />
+    public int GetVertexDegree(TVertex vertex) =>
+        GetVertexIndegreeCore(vertex) + GetVertexOutdegreeCore(vertex);
+
+    int GetVertexIndegreeCore(TVertex vertex)
+    {
         if (m_ReverseAdjacencyList != null)
             return GetVertexOutdegreeCore(m_ReverseAdjacencyList, vertex);
         else
             return m_AdjacencyList.Count(kv => kv.Value?.Contains(vertex) == true);
     }
 
-    /// <inheritdoc />
-    public int GetVertexOutdegree(TVertex vertex) => GetVertexOutdegreeCore(m_AdjacencyList, vertex);
+    int GetVertexOutdegreeCore(TVertex vertex) => GetVertexOutdegreeCore(m_AdjacencyList, vertex);
 
     static int GetVertexOutdegreeCore(AssociativeArray<TVertex, AdjacencyRow?> adjacencyList, TVertex vertex) =>
         adjacencyList.TryGetValue(vertex, out var adjacencyRow) ?
             adjacencyRow?.Count ?? 0 :
             0;
-
-    /// <inheritdoc />
-    public int GetVertexDegree(TVertex vertex) =>
-        GetVertexIndegree(vertex) + GetVertexOutdegree(vertex);
 }